Revert "Feature/ControlSystem (#267)"
This reverts commit a430b0285f676038083294c2a9c3a8773e378688.
