+++ Core/Inc/can_handler.h
@@ -3,13 +3,6 @@
 
 #include "can.h"
 #include "cmsis_os.h"
-#include "control.h"
-#include "dti.h"
-
-typedef struct {
-	dti_t *mc;
-	control_t *calypso_states;
-} can_receive_t;
 
 /**
  * @brief Callback to be called when a message is received on CAN line 1.
