+++ Core/Src/can_handler.c
@@ -9,7 +9,6 @@
 #include "dti.h"
 #include "fault.h"
 #include "steeringio.h"
-#include "control.h"
 
 #define CAN_MSG_QUEUE_SIZE 50 /* messages */
 
@@ -23,10 +22,8 @@ static osMessageQueueId_t can_inbound_queue;
 can_t *can1;
 
 /* Relevant Info for Initializing CAN 1 */
-static uint32_t id_list[] = { DTI_CANID_ERPM,	   DTI_CANID_CURRENTS,
-			      BMS_DCL_MSG,	   BUTTON_CANID_IO,
-			      DIAL_CANID_IO,	   CONTROL_CANID_FANBATTBOX,
-			      CONTROL_CANID_RADFAN };
+static uint32_t id_list[] = { DTI_CANID_ERPM, DTI_CANID_CURRENTS, BMS_DCL_MSG,
+			      BUTTON_CANID_IO, DIAL_CANID_IO };
 
 void init_can1(CAN_HandleTypeDef *hcan)
 {
@@ -38,14 +35,9 @@ void init_can1(CAN_HandleTypeDef *hcan)
 
 	can1->hcan = hcan;
 
-	uint32_t id_list_size_four1[4] = { id_list[0], id_list[1], id_list[2],
-					   id_list[3] };
-
-	uint32_t id_list_size_four2[4] = { id_list[4], id_list[5], id_list[6],
-					   id_list[6] };
-
-	assert(!can_add_filter(can1, id_list_size_four1));
-	assert(!can_add_filter(can1, id_list_size_four2));
+	uint32_t id_list_size_four[4] = { id_list[0], id_list[1], id_list[2],
+					  id_list[3] };
+	assert(!can_add_filter(can1, id_list_size_four));
 	assert(!can_init(can1));
 
 	can_outbound_queue =
@@ -140,8 +132,8 @@ const osThreadAttr_t can_receive_attributes = {
 
 void vCanReceive(void *pv_params)
 {
-	can_receive_t *can_receive = (can_receive_t *)pv_params;
-	assert(can_receive);
+	dti_t *mc = (dti_t *)pv_params;
+	assert(mc);
 
 	can_msg_t msg;
 
@@ -153,10 +145,7 @@ void vCanReceive(void *pv_params)
 			switch (msg.id) {
 			/* Messages Relevant to Motor Controller */
 			case DTI_CANID_ERPM:
-				dti_record_rpm(can_receive->mc, msg);
-				break;
-			case DTI_CANID_TEMPS_FAULT:
-				dti_record_temp(can_receive->mc, msg);
+				dti_record_rpm(mc, msg);
 				break;
 			case BMS_DCL_MSG:
 				handle_dcl_msg();
@@ -167,16 +156,6 @@ void vCanReceive(void *pv_params)
 			case DIAL_CANID_IO:
 				dial_update(msg);
 				break;
-			case CONTROL_CANID_FANBATTBOX:
-				control_fanbattbox_record(
-					can_receive->calypso_states, msg);
-				break;
-			case CONTROL_CANID_PUMP:
-				control_pump_record(can_receive->calypso_states,
-						    msg);
-			case CONTROL_CANID_RADFAN:
-				control_radfan_record(
-					can_receive->calypso_states, msg);
 			default:
 				break;
 			}
