+++ Core/Src/dti.c
@@ -17,17 +17,12 @@
 
 #include "emrax.h"
 #include "bms.h"
-#include "can_handler.h"
 
 #define CAN_QUEUE_SIZE 5 /* messages */
 #define SAMPLES	       20
 static osMutexAttr_t dti_mutex_attributes;
 
-static uint16_t mph = 0;
-
-static uint16_t motorTemp = 0;
-
-static uint16_t controllerTemp = 0;
+static uint8_t mph = 0;
 
 dti_t *dti_init()
 {
@@ -287,33 +282,6 @@ void dti_record_rpm(dti_t *mc, can_msg_t msg)
 	mc->rpm = rpm;
 	osMutexRelease(*mc->mutex);
 	mph = dti_get_mph(mc);
-}
-
-void dti_record_temp(dti_t *mc, can_msg_t msg)
-{
-	uint16_t controllerTemp = (msg.data[0] << 8) + (msg.data[1]);
-	uint16_t motorTemp = (msg.data[2] << 8) + (msg.data[3]);
-
-	controllerTemp /= 10;
-	motorTemp /= 10;
-
-	osMutexAcquire(*mc->mutex, osWaitForever);
-	mc->contr_temp = controllerTemp;
-	mc->motor_temp = motorTemp;
-	osMutexRelease(*mc->mutex);
-
-	controllerTemp = controllerTemp;
-	motorTemp = motorTemp;
-}
-
-uint16_t dti_get_motor_temp()
-{
-	return motorTemp;
-}
-
-uint16_t dti_get_controller_temp()
-{
-	return controllerTemp;
 }
 
 uint8_t get_mph()
