+++ Core/Src/main.c
@@ -37,7 +37,6 @@
 #include "dti.h"
 #include "steeringio.h"
 #include "pedals.h"
-#include "control.h"
 #include "monitor.h"
 #include "state_machine.h"
 /* USER CODE END Includes */
@@ -251,25 +250,10 @@ int main(void)
   // shutdown_monitor_handle = osThreadNew(vShutdownMonitor, pdu, &shutdown_monitor_attributes);
   // assert(shutdown_monitor_handle);
 
-  /* Control File Thread */
-  control_args_t *control_args = malloc(sizeof(control_args_t));
-  control_args->pdu = pdu;
-  control_args->control = malloc(sizeof(control_t));
-  control_args->calypso_states = malloc(sizeof(control_t));
-  control_args->control->fanBattBoxState = 0;
-  control_args->control->pumpState0 = 0;
-  control_args->control->pumpState1 = 0;
-  control_handle = osThreadNew(vControl, control_args, &control_attributes);
-  assert(control_handle);
-
   /* Messaging */
   can_dispatch_handle = osThreadNew(vCanDispatch, &hcan1, &can_dispatch_attributes);
   assert(can_dispatch_handle);
-
-  can_receive_t *can_receive = malloc(sizeof(can_receive));
-  can_receive->mc = mc;
-  can_receive->calypso_states = control_args->calypso_states;
-  can_receive_thread = osThreadNew(vCanReceive, can_receive, &can_receive_attributes);
+  can_receive_thread = osThreadNew(vCanReceive, mc, &can_receive_attributes);
   assert(can_receive_thread);
 
   /* Control Logic */
@@ -292,7 +276,6 @@ int main(void)
   sm_args->mpu = mpu;
   sm_director_handle = osThreadNew(vStateMachineDirector, sm_args, &sm_director_attributes);
   assert(sm_director_handle);
-
   /* USER CODE END RTOS_THREADS */
 
   /* USER CODE BEGIN RTOS_EVENTS */
@@ -869,7 +852,7 @@ void StartDefaultTask(void *argument)
     // refresh the external watchdog so the car doesnt fault
     pet_watchdog(mpu);
 
-	
+    
     /* Send NERO state data continuously */
     send_git_version_message();
     osDelay(500);
