+++ Core/Src/pdu.c
@@ -308,6 +308,16 @@ static int8_t write_ctrl(pdu_t *pdu, bool state, uint8_t pin, uint8_t reg)
 	return 0;
 }
 
+int8_t write_pump_0(pdu_t *pdu, bool state)
+{
+	return write_ctrl(pdu, state, PIN_PUMP_CTRL0, PCA_OUTPUT_0_REG);
+}
+
+int8_t write_pump_1(pdu_t *pdu, bool state)
+{
+	return write_ctrl(pdu, state, PIN_PUMP_CTRL1, PCA_OUTPUT_0_REG);
+}
+
 int8_t write_24V_12V_buck(pdu_t *pdu, bool state)
 {
 	return write_ctrl(pdu, state, PIN_BUCK_CTRL, PCA_OUTPUT_0_REG);
@@ -318,9 +328,24 @@ int8_t write_brakelight(pdu_t *pdu, bool state)
 	return write_ctrl(pdu, state, PIN_BRKLIGHT_CTRL, PCA_OUTPUT_0_REG);
 }
 
+int8_t write_fan_battbox(pdu_t *pdu, bool state)
+{
+	return write_ctrl(pdu, state, PIN_FANBATTBOX_CTRL, PCA_OUTPUT_0_REG);
+}
+
 int8_t write_rtds(pdu_t *pdu, bool state)
 {
 	return write_ctrl(pdu, state, PIN_RTD_CTRL, PCA_OUTPUT_1_REG);
+}
+
+int8_t write_radfan_0(pdu_t *pdu, bool state)
+{
+	return -1; // Replace with actual stuff when PDU Radfan CTRL is added to board
+}
+
+int8_t write_radfan_1(pdu_t *pdu, bool state)
+{
+	return -1; // Replace with actual stuff when PDU Radfan CTRL is added to board
 }
 
 /* Read Pump Sensors ADC DMA */
