+++ Core/Src/state_machine.c
@@ -8,7 +8,6 @@
 #include "cerb_utils.h"
 #include "monitor.h"
 #include "pedals.h"
-#include "can_handler.h"
 
 #define STATE_TRANS_QUEUE_SIZE 4
